redirect + delete and fetch in parent component
