+++ src/components/List/List.jsx
@@ -1,35 +1,28 @@
 import React, { Component } from "react";
 import Card from "./Card";
-const axios = require('axios');
+
 
 class List extends Component {
-  constructor(props) {
-    super(props)
-    
-    this.state = {
-      allTheNews: [],
-    }
-  }
   
   // Obtenemos la data del NY Times 
-  async componentDidMount() {
-    
-    const getNewsData = await axios.get(`https://api.nytimes.com/svc/topstories/v2/home.json?api-key=${process.env.REACT_APP_NYKEY}`);
-    const nyTimesNews = getNewsData.data.results.slice(0, 5)
-    const myNews = nyTimesNews.concat(this.props.createdNews)
-    this.setState({
-      allTheNews: myNews
-    })
-  }
+  // async componentDidMount() {
+  //   const getNewsData = await axios.get(`https://api.nytimes.com/svc/topstories/v2/home.json?api-key=${process.env.REACT_APP_NYKEY}`);
+  //   const nyTimesNews = getNewsData.data.results.slice(0, 5)
+  //   const myNews = nyTimesNews.concat(this.props.createdNews)
+  //   this.setState({
+  //     allTheNews: myNews
+  //   })
+  // }
 
   paintNews = () => {
-    return this.state.allTheNews.map((news, i) =>  <Card info={news} delete={() => this.deleteOneNews(i)} key={i}/>)
+    return this.props.allTheNews.map((news, i) =>  <Card info={news} delete={() => this.props.deleteOneNews(i)} key={i}/>)
   }
 
-  deleteOneNews = (i) => {
-    const news = this.state.allTheNews.filter((news, j) => j !== i)
-    this.setState({ allTheNews: news })
-  }
+  // deleteOneNews = (i) => {
+  //   const news = this.state.allTheNews.filter((news, j) => j !== i)
+  //   this.setState({ allTheNews: news })
+  //   this.props.createdNews({createdNews: news})
+  // }
 
   render() {
     return (<div>
